--- project/src/main/java/com/project/main/services/AddSkillsService.java
@@ -1,24 +1,0 @@
-package com.project.main.services;
-
-import com.project.main.models.Skill;
-import com.project.main.models.SkillForm;
-import com.project.main.models.User;
-import com.project.main.repositories.SkillRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class AddSkillsService {
-
-  @Autowired
-  private SkillRepository skillRepository;
-
-  public void addSkills(SkillForm[] skillsForm, User user) {
-
-    Skill[] skills = SkillForm.toSkills(skillsForm, user);
-    for (Skill skill : skills) {
-      skillRepository.save(skill);
-    }
-  }
-
-}
