+++ project/src/main/java/com/project/main/services/SkillService.java
@@ -1,11 +1,11 @@
 package com.project.main.services;
 
 import java.util.List;
-
 import com.project.main.models.Candidate;
 import com.project.main.models.Skill;
+import com.project.main.models.SkillForm;
+import com.project.main.models.User;
 import com.project.main.repositories.SkillRepository;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,6 +17,14 @@ public class SkillService {
 
   public List<Skill> getSkills(Candidate candidate) {
     return skillRepository.findByCandidateId(candidate.getId());
+  }
+
+  public void addSkills(SkillForm[] skillsForm, User user) {
+
+    Skill[] skills = SkillForm.toSkills(skillsForm, user);
+    for (Skill skill : skills) {
+      skillRepository.save(skill);
+    }
   }
 
 }
